// src/utils/types/problem.ts
import type { StaticImageData } from "next/image";

export type Example = {
	id: number;
	inputText: string;
	outputText: string;
	explanation?: string;
	img?: StaticImageData;
};

// local problem data
export type Problem = {
	id: string;
	title: string;
	problemStatement: string;
	examples: Example[];
	constraints: string;
	order: number;
	starterCode: string;
	handlerFunction: ((fn: any) => boolean) | string;
	starterFunctionName: string;
};

export type DBProblem = {
	id: string;
	title: string;
	category: string;
	difficulty: string;
	likes: number;
	dislikes: number;
	order: number;
	videoId?: string;
	link?: string;
};